refactor(app): extract bookmark orchestrators to bookmark_actions (phase 3c)
Move set_bookmark, jump_to_bookmark, clear_bookmark, list_bookmarks
out of render.rs into a new app/bookmark_actions.rs sibling. ~88 lines.

These methods were already going through the BookmarkState API
extracted in phase 2b; this commit is purely structural.

render.rs goes from 3602 -> 3514 lines (-88). 2026 tests still pass.
